pyproject.toml
[project]
name = "autoTable"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "gouzil", email = "530971494@qq.com"},
]
dependencies = [
    "loguru>=0.7.2",
    "typer>=0.9.0",
    "PyGithub>=2.1.1",
    "rich>=13.7.0",
    "mistletoe>=1.2.1",
    "pytest"
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[project.scripts]
autotable = "autotable.__main__:app"

[tool.pdm.scripts]
start = "autotable issue-update 1 2 3"
save = "autotable issue-backup"
test.cmd = "pytest"
test.help = "Launch pytest"

[tool.ruff]
line-length = 120
src = ["autotable", "tests"]
target-version = "py310"

[tool.ruff.lint]
select = [
  "Q",
  "F",
  "SIM",
  "UP",
  "FA",   # flake8-annotations
  "I",    # isort
  "B",    # flake8-bugbear
  "C4",   # flake8-comprehensions
  "PGH",  # pygrep-hooks
  "RUF",  # ruff
  "E",    # pycodestyle
  "W",    # pycodestyle
  "YTT",  # flake8-2020
]

[tool.ruff.lint.isort]
known-first-party = ["autotable"]
required-imports = ["from __future__ import annotations"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401", "I002"]

[tool.pyright]
include = ["autotable"]

[tool.pdm.dev-dependencies]
test = [
    "xdoctest",
    "pytest",
    "pytest-cov",
    "coverage",
]

[tool.pdm]
package-type = "library"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
